test: Refactor tests
