+++ test/unit/agents.test.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import "mocha";
 
-import { getDemoMetadata, log } from "../test-utils";
+import { getDemoMetadata, log, Dummy, makeAgents } from "../test-utils";
 import { Game } from "../../src/api";
 import {
     Agent,
@@ -15,12 +15,6 @@ import {
 } from "../../src/events/event-record";
 import { on } from "../../src/events";
 import { buildGameInstance } from "../../src/state";
-
-class Dummy extends Agent {
-    constructor(public name: string, public health: number) {
-        super();
-    }
-}
 
 class Parent extends Agent {
     constructor(public child: Dummy) {
@@ -38,19 +32,6 @@ class MultiParent extends Agent {
     constructor(public children: Dummy[]) {
         super();
     }
-}
-
-function makeAgents(startFrom: number, amount: number) {
-    let num = startFrom;
-    let i = amount;
-    const arr = [];
-
-    while (i-- > 0) {
-        arr.push(new Dummy(`D${num}`, num));
-        num++;
-    }
-
-    return arr;
 }
 
 const MD = getDemoMetadata();
@@ -2651,168 +2632,6 @@ describe("Agents", function() {
             expect(StaticAgentRegistry.hasAgent(2)).to.be.false;
             expect(StaticAgentRegistry.hasAgent(3)).to.be.false;
             expect(StaticAgentRegistry.getNextAvailableId()).to.equal(1);
-        });
-    });
-
-    describe("Reverting", function() {
-        it("Reverting the effects of many events back to the first one", function() {
-            const init = on("INIT", game => {
-                game.state.foo = "Hello, world!";
-            });
-
-            const mod = (num: number) =>
-                on(`MOD ${num}`, game => {
-                    game.state.foo += `-${num}`;
-                    game.state[num] = `Yo ${num}`;
-                });
-
-            Game.init(MD);
-
-            let myGame = buildGameInstance({ trackAgentChanges: true });
-            init(myGame);
-
-            for (let x = 0; x < 10; x++) {
-                mod(x)(myGame);
-            }
-
-            expect(myGame.state.foo).to.equal(
-                "Hello, world!-0-1-2-3-4-5-6-7-8-9"
-            );
-            expect(myGame.state[0]).to.equal("Yo 0");
-            expect(myGame.state[5]).to.equal("Yo 5");
-            expect(myGame.state[9]).to.equal("Yo 9");
-
-            myGame = myGame.revert(1);
-
-            expect(myGame.state.foo).to.equal("Hello, world!");
-            expect(myGame.state[0]).to.be.undefined;
-            expect(myGame.state[5]).to.be.undefined;
-            expect(myGame.state[9]).to.be.undefined;
-        });
-
-        it("Reverting the effects of many events to different points in time", function() {
-            const init = on("INIT", game => {
-                game.state.foo = "Hello, world!";
-            });
-
-            const mod = (num: number) =>
-                on(`MOD ${num}`, game => {
-                    game.state.foo += `-${num}`;
-                    game.state[num] = `Yo ${num}`;
-                });
-
-            Game.init(MD);
-
-            const myGame = buildGameInstance({ trackAgentChanges: true });
-            init(myGame);
-
-            for (let x = 0; x < 10; x++) {
-                mod(x)(myGame);
-            }
-
-            expect(myGame.state.foo).to.equal(
-                "Hello, world!-0-1-2-3-4-5-6-7-8-9"
-            );
-            expect(myGame.state[0]).to.equal("Yo 0");
-            expect(myGame.state[5]).to.equal("Yo 5");
-            expect(myGame.state[9]).to.equal("Yo 9");
-
-            const g1 = myGame.revert(8);
-
-            expect(g1.state.foo).to.equal("Hello, world!-0-1-2-3-4-5-6");
-            expect(g1.state[0]).to.equal("Yo 0");
-            expect(g1.state[5]).to.equal("Yo 5");
-            expect(g1.state[9]).to.be.undefined;
-
-            const g2 = myGame.revert(6);
-
-            expect(g2.state.foo).to.equal("Hello, world!-0-1-2-3-4");
-            expect(g2.state[0]).to.equal("Yo 0");
-            expect(g2.state[5]).to.be.undefined;
-            expect(g2.state[9]).to.be.undefined;
-
-            const g3 = myGame.revert(4);
-
-            expect(g3.state.foo).to.equal("Hello, world!-0-1-2");
-            expect(g3.state[0]).to.equal("Yo 0");
-            expect(g3.state[5]).to.be.undefined;
-            expect(g3.state[9]).to.be.undefined;
-
-            const g4 = myGame.revert(1);
-
-            expect(g4.state.foo).to.equal("Hello, world!");
-            expect(g4.state[0]).to.be.undefined;
-            expect(g4.state[5]).to.be.undefined;
-            expect(g4.state[9]).to.be.undefined;
-        });
-
-        it("Reverting to before an agent was registered", function() {
-            const init = on("INIT", game => {
-                game.state.foo = "Hello, world!";
-                game.state.dummy = new Dummy("Lars", 10);
-            });
-
-            Game.init(MD);
-
-            let myGame = buildGameInstance();
-            init(myGame);
-            myGame = myGame.revert(0);
-
-            expect(myGame.state.foo).to.be.undefined;
-            expect(myGame.state.dummy).to.be.undefined;
-            expect(myGame.agents.getAgentProperty(1, "name")).to.be.undefined;
-            expect(myGame.agents.getAgentProperty(1, "health")).to.be.undefined;
-        });
-
-        it("Reverting changes to static agents", function() {
-            const staticDummy = new Dummy("Lars", 15);
-
-            Game.init(MD);
-
-            let myGame = buildGameInstance();
-            on("FUNC", game => {
-                const dummy = game.using(staticDummy);
-                dummy.name = "Jimbo";
-                dummy["bippity"] = "boppity";
-            })(myGame);
-
-            expect(myGame.agents.getAgentProperty(1, "name")).to.equal("Jimbo");
-            expect(myGame.agents.getAgentProperty(1, "health")).to.equal(15);
-            expect(myGame.agents.getAgentProperty(1, "bippity")).to.equal(
-                "boppity"
-            );
-
-            myGame = myGame.revert();
-
-            expect(myGame.agents.getAgentProperty(1, "name")).to.equal("Lars");
-            expect(myGame.agents.getAgentProperty(1, "health")).to.equal(15);
-            expect(myGame.agents.getAgentProperty(1, "bippity")).to.be
-                .undefined;
-        });
-
-        it("Reverting changes to an agent array through the Game API", function() {
-            Game.init(MD);
-            const myGame = buildGameInstance();
-            myGame.state.agents = makeAgents(0, 6);
-
-            expect(myGame.state.agents.length).to.equal(6);
-            expect(myGame.state.agents[0].name).to.equal("D0");
-            expect(myGame.state.agents[5].name).to.equal("D5");
-
-            on("REVERSE", game => {
-                const agents = myGame.state.agents as Dummy[];
-                agents.reverse();
-            })(myGame);
-
-            expect(myGame.state.agents.length).to.equal(6);
-            expect(myGame.state.agents[0].name).to.equal("D5");
-            expect(myGame.state.agents[5].name).to.equal("D0");
-
-            const res = Game.postUndoCommand(myGame).instance;
-
-            expect(res.state.agents.length).to.equal(6);
-            expect(res.state.agents[0].name).to.equal("D0");
-            expect(res.state.agents[5].name).to.equal("D5");
         });
     });
